+++ mayan/apps/document_states/workflow_actions.py
@@ -83,7 +83,7 @@ class DocumentPropertiesEditAction(WorkflowAction):
             document.save()
 
 
-class HTTPPostAction(WorkflowAction):
+class HTTPAction(WorkflowAction):
     fields = {
         'url': {
             'label': _('URL'),
@@ -146,6 +146,20 @@ class HTTPPostAction(WorkflowAction):
                     'and "comment" attributes.'
                 ), 'max_length': 192, 'required': False
             },
+        }, 'method': {
+            'label': _('Method'),
+            'class': 'django.forms.CharField', 'kwargs': {
+                'help_text': _(
+                    'The HTTP method to use for the request. Typical choices '
+                    'are OPTIONS, HEAD, POST, GET, PUT, PATCH, DELETE. '
+                    'Can be a static value or a template that returns the '
+                    'method text. Templates receive the workflow log entry '
+                    'instance as part of their context via the '
+                    'variable "entry_log". The "entry_log" in turn '
+                    'provides the "workflow_instance", "datetime", '
+                    '"transition", "user", and "comment" attributes.'
+                ), 'required': True
+            }
         }, 'headers': {
             'label': _('Headers'),
             'class': 'django.forms.CharField', 'kwargs': {
@@ -162,9 +176,12 @@ class HTTPPostAction(WorkflowAction):
         }
     }
     field_order = (
-        'url', 'username', 'password', 'headers', 'timeout', 'payload'
+        'url', 'username', 'password', 'headers', 'timeout', 'verb', 'payload'
     )
-    label = _('Perform a POST request')
+    label = _('Perform an HTTP request')
+    previous_dotted_paths = (
+        'mayan.apps.document_states.workflow_actions.HTTPPostAction',
+    )
     widgets = {
         'payload': {
             'class': 'django.forms.widgets.Textarea', 'kwargs': {
@@ -216,12 +233,13 @@ class HTTPPostAction(WorkflowAction):
         return result
 
     def execute(self, context):
+        headers = self.render_load(field_name='headers', context=context)
+        method = self.render(field_name='method', context=context)
+        password = self.render(field_name='password', context=context)
+        payload = self.render_load(field_name='payload', context=context)
+        timeout = self.render(field_name='timeout', context=context)
         url = self.render(field_name='url', context=context)
         username = self.render(field_name='username', context=context)
-        password = self.render(field_name='password', context=context)
-        timeout = self.render(field_name='timeout', context=context)
-        headers = self.render_load(field_name='headers', context=context)
-        payload = self.render_load(field_name='payload', context=context)
 
         if '.' in timeout:
             timeout = float(timeout)
@@ -236,7 +254,7 @@ class HTTPPostAction(WorkflowAction):
                 username=username, password=password
             )
 
-        requests.post(
-            url=url, json=payload, timeout=timeout,
+        requests.request(
+            method=method, url=url, json=payload, timeout=timeout,
             auth=authentication, headers=headers
         )